Removed paydaymods
Added a check for .host value for updates to be valid. Avoiding erroring due to old mods.
Changed localization for English localizatoin mentioning paydaymods (will need to be done for other langauges as well).
Added a check for patch notes to avoid displaying them if the mod did not provide them.
Changed readme to match SuperBLT.
Disabled RetrieveDependencies function as it does nothing at the moment.