Fix crash when trying to convert log name to number
If for some reason a textfile that contains at least 3 underscores is saved to the logs folder but isn't using YY_MM_DD_, the game will crash upon launch because LogNameToNumber is calling TimestampToEpoch with one or more nil values for year/month/day.
It is better to use a less loose filename matching in this case, that also allows other mods to write to logs without causing trouble.