Fix DelayedCalls undefined behavior
Adding a DelayedCall inside another DelayedCall leads to undefined behavior, since the _calls table is currently being iterated on.
Saving the _calls table to a local, then clearing it and iterating over the local makes sure that calls added during iteration will be saved properly.
To allow for call removal with the new system, _remove_queue holds all calls that should be removed and is checked for every call during iteration.