Sort mod manager alphabetical and slightly mute disabled mods
Sorting by priority is not really useful for the average user and just makes finding mods harder overall since the order is kind of random when you have a lot of mods with default priority.
Libraries will still be sorted first and disabled mods will have their alpha reduced to make it easier to see that they are disabled.