Re-implement dependencies
Dependencies can now be listed in the mod.txt again and SuperBLT will try to download them.
This can be done in two ways, by providing a direct download link or by providing a link to a meta file (the same that is used for auto updates of a mod).
In addition to the old behavior where only mods that had auto updating could be added as dependencies simple dependencies can now be added by adding a key value pair of ``"Mod Name" : "mod.download.url"`` which is equal to specifying ``"Mod Name" : { "download_url" : "mod.download.url" }``.
When linking to meta files, one would use this format ``"mod-ident" : { "meta" : "mod.meta.url" }``.
Additional parameters can be added to the dependency data such as display name or install directory.